refactor: move components related to Article to src/components directory
